Setup backend with heroku (#5)
* Setup heroku with the app backend
