--- backend/data.json
@@ -1,7 +1,0 @@
-{"test_data":
-    {
-        "event1": 100,
-        "event2": 200,
-        "event3": 300
-    }
-}
